net: change CNetAddr::ip to have flexible size
Before this change `CNetAddr::ip` was a fixed-size array of 16 bytes,
not being able to store larger addresses (e.g. TORv3) and inefficiently
encoded smaller addresses as 16-byte IPv6 addresses.

Change it to be `std::vector`, so that it can hold larger addresses and
also do not encode smaller addresses into 16 bytes anymore. So the IPv4
address `1.2.3.4` is now encoded as `01020304` (4 bytes) instead of
`00000000000000000000FFFF01020304` (16 bytes).

Rename `CNetAddr::ip` to `CNetAddr::m_addr` because it is not an "IP" or
"IP address" (TOR addresses are not IP addresses).

In order to preserve backward compatibility with serialization (where
e.g. `1.2.3.4` is serialized as `00000000000000000000FFFF01020304`)
introduce `CNetAddr` dedicated legacy serialize/unserialize methods.

Adjust `CSubNet` accordingly. Still use `CSubNet::netmask[]` of fixed 16
bytes, but use the first 4 for IPv4 (not the last 4). Do not accept
invalid netmasks that have 0-bits followed by 1-bits and only allow
subnetting for IPv4 and IPv6 because it is easier that way.

Co-authored-by: Carl Dong <contact@carldong.me>