refactor: Prevent double lock in CAddrMan::size()
Cache the value of `vRandom.size()` in an atomic variable `m_size` and
read it without a mutex protection.

Modifications of `m_size` are done in the same places where `vRandom` is
modified which are protected by a mutex. This prevents the values of
`vRandom.size()` and the new variable from going out of sync.

Readers that call `CAddrMan::size()` could read the new size earlier,
compared to before this patch - while another thread is still modifying
`CAddrMan` and has not released the mutex yet.