net: don't extra bind for Tor if binds are restricted
The semantic of `-bind` is to restrict the binding only to some address.
If not specified, then the user does not care and we bind to `0.0.0.0`.
If specified then we should honor the restriction and bind only to the
specified address.

Before this change, if no `-bind` is given then we would bind to
`0.0.0.0` and to `127.0.0.1:8334` (incoming Tor) which is ok - the user
does not care to restrict the binding. However:

* If only `-bind=addr` is given (without `-bind=...=onion`) then we
  would bind to `addr` _and_ to `127.0.0.1:8334` in addition which may
  be unexpected assuming the perception of `-bind=addr` is
  "bind _only_ to `addr`".

* If only `-bind=addr=onion` is given (without ordinary `-bind=`)
  then we would bind to `addr` _and_ to `0.0.0.0` in addition.

Change the above two cases to not do the additional bind. If `-bind`
is given then only bind to the specified address:

* If only `-bind=addr` is given (without `-bind=...=onion`) then bind to
  `addr` (only). If we are creating tor hidden service then use `addr`
  as target. Same behavior as before
  https://github.com/bitcoin/bitcoin/pull/19991.

* If only `-bind=addr=onion` is given (without ordinary `-bind=`)
  then bind to `addr` (only) and consider incoming connections as Tor
  and do not advertise `addr`. I.e. a Tor-only node.