netbase: log netbase messages with category and debug level
and change a few of the NET category logging messages to PROXY when the message
relates to SOCKS5, per the following log category descriptions in
https://github.com/jonatack/bitcoin-development/blob/master/logging.md
which is an updated version of this Bitcoin Stack Exchange answer in 2017 by Andrew Chow:
https://bitcoin.stackexchange.com/questions/66892/what-are-the-debug-categories/66895

- net: All messages related to communicating with other nodes on the network,
  including what P2P messages were sent and received and to whom and other
  information about the network messages.

- proxy: Messages about using a SOCKS5 proxy and its authentication.

Before this commit, BCLog::PROXY was essentially unused (only one log message).

and reduce some unconditional logging that could be peer-provoked.

While here, improve the clarity of a few of the touched messages and use
consistent capitalization and naming, e.g. "Socks5()" vs "SOCKS5"