net_processing: replace LogPrintf messages with category and debug level
using LogPrintLevel(), thereby removing all LogPrintf messages in this file.

Reduce some of the unconditional logging that could be peer-provoked
(disconnection for old chain, stalling, timeout, but not user-selected
cases like manual/noban/forcerelay peers), by moving it to the Debug level.

While here, improve the clarity of a few of the touched messages and use
consistent capitalization.