fixup! cmake: Add `Coverage` and `CoverageFuzz` scripts
Use an option for enabling coverage instead of a build target

Enabling code coverage boils down to adding extra compiler options,
like the WERROR option. This can be done for any build target and
regardless of the build target, thus switch from using a build target
to a -DCOVERAGE=ON|OFF option.

Also use the native Clang options when the compiler is Clang, as per
https://clang.llvm.org/docs/SourceBasedCodeCoverage.html#compiling-with-coverage-enabled