net: simplify deletion of objects from CConnman::m_nodes
Before this change unneeded `CNode`s were scheduled for deletion by
putting them in `CConnman::m_nodes_disconnected` and were deleted after
they were no longer referenced, whereas new references can only be added
to `CConnman::m_nodes` and never to `CConnman::m_nodes_disconnected`.

This somewhat duplicates the logic of `shared_ptr`. Remova all of that
and make use of the fact that `shared_ptr` will destroy the object when
it is no longer referenced.

Deleting consists of calling `PeerManager::FinalizeNode()` and deleting
the `CNode`. So call `FinalizeNode()` from `~CNode()`. This way is it
sure that `PeerManager` will always be informed that a `CNode` is being
disposed.

`CConnman::NodesSnapshot` is not necessary anymore because its
functionality is covered by the standard library - making a copy of
a `vector` of `shared_ptr`s increments the reference count of each
member of the `vector`.